fix: progress data races (#17)
Packfile writers are created and the index gets built in a separate
goroutine. A data race occurred when setting up the progress
collector. I plumbed the progress collector into the `newPackWrite`
funtion outright. That way, the call to `buildIndex` could take
it. Furthermore, using a mutex to serialize reporting for receiving
objects and resolving deltas.